requirements.txt
# Google Maps Scraper — Python Dependencies
# Author    : Soubarna Karmakar
# Copyright : © 2025 Soubarna Karmakar. All rights reserved.
#
# Install: pip install -r requirements.txt

# ── Web framework + ASGI server ───────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
python-multipart>=0.0.9

# ── Desktop window (wraps the React UI — no browser bar) ─────────────────────
pywebview>=5.0.0

# ── Browser automation ────────────────────────────────────────────────────────
playwright>=1.43.0

# ── Excel export ──────────────────────────────────────────────────────────────
openpyxl>=3.1.2

# ── Email extraction from websites (optional) ─────────────────────────────────
requests>=2.31.0
beautifulsoup4>=4.12.3
lxml>=5.2.0
